% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winners-data.R
\docType{data}
\name{winners}
\alias{winners}
\title{Tour de France winners data}
\format{A \code{dataframe} with 106 rows and 17 columns.}
\usage{
data(winners)
}
\description{
Tour de France winners data
}
\examples{
data(winners)
head(winners)
}
\keyword{datasets}
